backend/requirement.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt[crypto]==5.5.1
django-cors-headers==4.7.0
python-decouple==3.8
Pillow==11.1.0
psycopg2-binary==2.9.10
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
whitenoise==6.9.0